1. Remove schedule_preemptive functions which are not part of the design 2. Implement the wait_allocation function on both sides. The client    would try until the timeout expires or the scheduler returns Some(resource) 3. Fix the return type of the wait_preemptive and list_allocations    functions 4. Implement the release function that returns Result<(), Error>    although the design states It returns nothing, the client json-rpc    library requires and sized type as a return. 5. Added test for the wait_allocation and release functions.

+++ common/src/resource.rs
@@ -15,5 +15,5 @@ pub struct ResourceReq {
 #[derive(Clone, Debug, serde::Serialize, serde::Deserialize)]
 pub struct ResourceAlloc {
     pub resource: ResourceReq,
-    pub resource_id: usize,
+    pub resource_id: u32,
 }
